Move GetCurrentStorageOSCluster to new pkg
GetCurrentStorageOSCluster() is useful at various places. Move it to a
separate package that can be imported and used easily.
