+++ internal/pkg/storageoscluster/doc.go
@@ -1,0 +1,2 @@
+// Package storageoscluster provides helpers related to StorageOSCluster object.
+package storageoscluster
